# HDDS-12349. Speed up some HDDS integration tests