--- a/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/hdds/scm/container/metrics/TestSCMContainerManagerMetrics.java
+++ b/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/hdds/scm/container/metrics/TestSCMContainerManagerMetrics.java
@@ -121,7 +121,7 @@ public abstract class TestSCMContainerManagerMetrics implements NonHATests.TestC
     assertEquals(getLongCounter("NumSuccessfulDeleteContainers",
         metrics), numSuccessfulDeleteContainers);
     assertEquals(getLongCounter("NumFailureDeleteContainers",
-        metrics), ++numFailureDeleteContainers);
+        metrics), numFailureDeleteContainers + 1);
 
     long currentValue = getLongCounter("NumListContainerOps", metrics);
     containerManager.getContainers(