## What changes were proposed in this pull request?

- Convert some HDDS integration tests to use shared cluster.
- Extract safe test case from `TestStorageContainerManagerHA` into a separate class, to be used with shared cluster.  Some trivial code cleanup in this test class, to avoid duplicating the problems.
- Move safe test cases from `TestQueryNode` to `TestContainerOperations` (which also tests `ScmClient`).
- Remove flaky tag from `TestDecommissionAndMaintenance`, passed 10x10 runs twice ([1](https://github.com/adoroszlai/ozone/actions/runs/13352899519), [2](https://github.com/adoroszlai/ozone/actions/runs/13358224366)).  This does not result in speed-up, but reduces run time of _integration (flaky)_, the slowest check currently.

https://issues.apache.org/jira/browse/HDDS-12349

## How was this patch tested?

Before:

```
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 158.4 s - in org.apache.hadoop.hdds.scm.TestStorageContainerManagerHA
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 35.973 s - in org.apache.hadoop.hdds.scm.TestXceiverClientManager
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 68.415 s - in org.apache.hadoop.hdds.scm.container.metrics.TestSCMContainerManagerMetrics
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 110.01 s - in org.apache.hadoop.hdds.scm.node.TestQueryNode
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 37.104 s - in org.apache.hadoop.ozone.TestContainerOperations
```

After:

```
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.832 s - in org.apache.ozone.test.HATests$StorageContainerManagerHAWithAllRunning
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 111.809 s - in org.apache.hadoop.hdds.scm.TestStorageContainerManagerHA
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.607 s - in org.apache.ozone.test.NonHATests$XceiverClientManager
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.82 s - in org.apache.ozone.test.NonHATests$SCMContainerManagerMetrics
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 44.172 s - in org.apache.hadoop.hdds.scm.node.TestQueryNode
Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.226 s - in org.apache.ozone.test.NonHATests$ContainerOperations
```

https://github.com/adoroszlai/ozone/actions/runs/13411975449